# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High performant JKJAV API Server",
        "title": "JKJAV API Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/list-ppv": {
            "get": {
                "description": "Get PPV slots by state",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List PPV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "list PPV by state. The only available option is PWTC",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.PPV"
                            }
                        }
                    }
                }
            }
        },
        "/submit": {
            "post": {
                "description": "Submit vaccine booking slot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Submit",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.PPV": {
            "type": "object",
            "properties": {
                "availability": {
                    "type": "integer",
                    "example": 10
                },
                "date": {
                    "type": "string",
                    "example": "20210516"
                },
                "location": {
                    "type": "string",
                    "example": "PWTC"
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Kuala Lumpur"
                },
                "date": {
                    "type": "string",
                    "example": "20210516"
                },
                "firstName": {
                    "type": "string",
                    "example": "Fadhil"
                },
                "lastName": {
                    "type": "string",
                    "example": "Yaacob"
                },
                "location": {
                    "type": "string",
                    "example": "PWTC"
                },
                "mysejahteraId": {
                    "type": "string",
                    "example": "900127015527"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "0123456789"
                }
            }
        }
    }
}